Split the fluid surface sampler out, and draw the overlay as water
ShipExteriorFluidSampler now owns resolving where the world's fluid surface
sits: vanilla-faithful per-corner column heights, their chunk-revision caches,
and the sample-point search. The overlay keeps only quad emission, and the
occluder depends on the sampler directly rather than reaching through it.

Restores the cache invalidation that was lost when this work moved off the old
branch. Nothing was bumping the chunk revisions, so a shoreline or a drained
pool kept reporting its old waterline forever; ship fluid was then built against
a surface that was no longer there.

The overlay is wired back in, but inside the translucent chunk layer rather than
after it. Iris resolves the terrain translucent shader to gbuffers_water only
while the phase is neither entities nor block entities, so drawing there hands
the quads to the pack's own water vertex shader instead of MOVING_BLOCK. That
also settles the vertex format: terrain programs declare their attributes
against DefaultVertexFormat.BLOCK, where the previous entity format would have
bound gbuffers_entities.
