+++ common/src/main/java/org/valkyrienskies/mod/common/fluid/client/ShipWaterPocketLiquidOverlay.java
@@ -4,8 +4,6 @@ import com.mojang.blaze3d.vertex.DefaultVertexFormat;
 import com.mojang.blaze3d.vertex.Tesselator;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.blaze3d.vertex.VertexFormat;
-import it.unimi.dsi.fastutil.longs.Long2LongOpenHashMap;
-import it.unimi.dsi.fastutil.longs.Long2ObjectOpenHashMap;
 import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.Comparator;
@@ -18,13 +16,10 @@ import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.RenderStateShard;
-import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.SectionPos;
 import net.minecraft.util.Mth;
 import net.minecraft.world.inventory.InventoryMenu;
-import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.block.DoorBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.Fluid;
@@ -48,22 +43,21 @@ import org.valkyrienskies.mod.util.FluidStateManager;
 /**
  * Geometry-based liquid overlay for submerged ship windows and openings.
  *
- * <p><b>Only the sampling half of this class is live.</b> Interior flood surfaces are drawn by
+ * <p>Closes off the world's water surface where it meets a ship's hull openings, so looking out
+ * through a submerged window shows water rather than the hole the hull cuts in the world's own
+ * fluid. This is the exterior waterline; interior flood surfaces belong to
  * {@link org.valkyrienskies.mod.common.fluid.FloodedVoxelRenderer}, which understands the streamed
- * partial fill amounts this overlay predates, so {@link #render} is deliberately not wired into the
- * render pipeline. What is still used is the exterior-water surface solver — the vanilla-faithful
- * column heights, their chunk-revision caches, and the boundary-mask helpers — which
- * {@link ShipPocketWorldWaterOccluder} and {@link ShipWaterPocketOverlayGeometry} depend on to know
- * where the waterline actually falls. Splitting that out is worthwhile follow-up; it is left intact
- * here so the surface math stays byte-for-byte what was tested.</p>
+ * partial fill amounts this class predates.</p>
+ *
+ * <p>Where the surface actually sits comes from {@link ShipExteriorFluidSampler}; which cells sit on
+ * the boundary comes from {@link ShipWaterPocketOverlayGeometry}. What is left here is the quad
+ * emission itself, drawn through {@link #OVERLAY_RENDER_TYPE} so a shaderpack shades it as water.</p>
  */
 public final class ShipWaterPocketLiquidOverlay {
 
     private ShipWaterPocketLiquidOverlay() {}
 
     private static final int MAX_SHIPS = 8;
-    private static final int MAX_FLUID_SURFACE_CACHE = 8192;
-    private static final int MAX_FLUID_SURFACE_POINT_CACHE = 16384;
     private static final double MAX_OVERLAY_SHIP_DISTANCE_BLOCKS_FALLBACK = 192.0;
     private static final double OVERLAY_SHIP_DISTANCE_MARGIN_BLOCKS = 48.0;
     private static final double OVERLAY_NEAR_VIEW_CULL_DISTANCE_BLOCKS = 24.0;
@@ -74,23 +68,6 @@ public final class ShipWaterPocketLiquidOverlay {
     private static final float FACE_SAMPLE_OFFSET = 0.15f;
     private static final float OVERLAY_UV_SCALE = 1.0f;
     private static final int FULL_BRIGHT = 0x00F000F0;
-    static final double SURFACE_EPS = 1.0E-5;
-    private static final double[][] FLUID_SAMPLE_OFFSETS = {
-        {0.0, 0.0, 0.0},
-        {0.0, -0.35, 0.0},
-        {0.35, 0.0, 0.0},
-        {-0.35, 0.0, 0.0},
-        {0.0, 0.0, 0.35},
-        {0.0, 0.0, -0.35},
-        {0.35, -0.35, 0.0},
-        {-0.35, -0.35, 0.0},
-        {0.0, -0.35, 0.35},
-        {0.0, -0.35, -0.35},
-        {0.35, 0.0, 0.35},
-        {0.35, 0.0, -0.35},
-        {-0.35, 0.0, 0.35},
-        {-0.35, 0.0, -0.35},
-    };
     private static final int SHAPE_SUBCELL_RES = 8;
     private static final int SHAPE_SUBCELL_COUNT = SHAPE_SUBCELL_RES * SHAPE_SUBCELL_RES * SHAPE_SUBCELL_RES;
     private static final int SHAPE_FACE_NEG_X = 0;
@@ -113,70 +90,8 @@ public final class ShipWaterPocketLiquidOverlay {
     private static final float[] CLIP_U1 = new float[6];
     private static final float[] CLIP_V1 = new float[6];
 
-    static final class FluidSurfaceSample {
-        final Fluid fluid;
-        final FluidState fluidState;
-        final BlockPos pos;
-        final double surfaceY;
 
-        private FluidSurfaceSample(final Fluid fluid, final FluidState fluidState, final BlockPos pos, final double surfaceY) {
-            this.fluid = fluid;
-            this.fluidState = fluidState;
-            this.pos = pos;
-            this.surfaceY = surfaceY;
-        }
-    }
 
-    /**
-     * The vanilla-equivalent per-corner surface heights of the fluid block at {@code pos} —
-     * position-independent within that block, so it's safe to cache and reuse for any fractional
-     * (x, z) query landing in the same block, unlike a single resolved Y. See {@link
-     * #computeVanillaColumnHeights} for how NW/NE/SW/SE are derived (a faithful port of vanilla's
-     * {@code LiquidBlockRenderer.calculateAverageHeight}), and {@link #surfaceYAt} for how a query
-     * point's fractional position within the block turns these into an actual world Y — matching
-     * what vanilla will actually rasterize instead of approximating it with a single block-uniform
-     * height (a real geometric mismatch that let world water render in front of our occluder in
-     * some spots and behind it in others, depending on local shoreline/flow topology).
-     */
-    private static final class ColumnHeights {
-        final Fluid fluid;
-        final FluidState fluidState;
-        final BlockPos pos;
-        final float nw;
-        final float ne;
-        final float sw;
-        final float se;
-
-        private ColumnHeights(final Fluid fluid, final FluidState fluidState, final BlockPos pos,
-            final float nw, final float ne, final float sw, final float se) {
-            this.fluid = fluid;
-            this.fluidState = fluidState;
-            this.pos = pos;
-            this.nw = nw;
-            this.ne = ne;
-            this.sw = sw;
-            this.se = se;
-        }
-
-        /** Bilinear blend of the four corner heights at the given world X/Z, matching vanilla's quad. */
-        private double surfaceYAt(final double worldX, final double worldZ) {
-            final double fx = Mth.clamp(worldX - pos.getX(), 0.0, 1.0);
-            final double fz = Mth.clamp(worldZ - pos.getZ(), 0.0, 1.0);
-            final double h = nw * (1.0 - fx) * (1.0 - fz) + ne * fx * (1.0 - fz)
-                + sw * (1.0 - fx) * fz + se * fx * fz;
-            return pos.getY() + h;
-        }
-    }
-
-    private static final class ChunkTrackedFluidSurface {
-        private final long chunkRevision;
-        private final ColumnHeights columnHeights;
-
-        private ChunkTrackedFluidSurface(final long chunkRevision, final ColumnHeights columnHeights) {
-            this.chunkRevision = chunkRevision;
-            this.columnHeights = columnHeights;
-        }
-    }
 
     private static final class OverlayFaceSample {
         private final TextureAtlasSprite sprite;
@@ -212,50 +127,25 @@ public final class ShipWaterPocketLiquidOverlay {
     }
 
     private static final Map<BlockState, OverlayShapeTemplate> SHAPE_TEMPLATE_CACHE = new HashMap<>();
-    private static final ColumnHeights COLUMN_HEIGHTS_MISS = new ColumnHeights(
-        net.minecraft.world.level.material.Fluids.EMPTY,
-        net.minecraft.world.level.material.Fluids.EMPTY.defaultFluidState(),
-        BlockPos.ZERO,
-        Float.NaN, Float.NaN, Float.NaN, Float.NaN
-    );
-    private static net.minecraft.client.multiplayer.ClientLevel lastSurfaceCacheLevel = null;
-    private static final Long2LongOpenHashMap EXTERIOR_FLUID_CHUNK_REVISIONS = new Long2LongOpenHashMap();
-    private static final Long2ObjectOpenHashMap<ChunkTrackedFluidSurface> FLUID_SURFACE_CACHE = new Long2ObjectOpenHashMap<>();
-    private static final Long2ObjectOpenHashMap<ChunkTrackedFluidSurface> FLUID_SURFACE_POINT_CACHE =
-        new Long2ObjectOpenHashMap<>();
     public static void clear() {
         SHAPE_TEMPLATE_CACHE.clear();
-        lastSurfaceCacheLevel = null;
         ShipWaterPocketOverlayGeometry.clear();
-        clearExteriorFluidCaches();
+        ShipExteriorFluidSampler.clear();
     }
 
-    public static void invalidateExteriorFluidChunk(final @Nullable net.minecraft.client.multiplayer.ClientLevel level, final int chunkX,
-        final int chunkZ) {
-        if (level == null) return;
-        if (lastSurfaceCacheLevel != null && lastSurfaceCacheLevel != level) {
-            clearExteriorFluidCaches();
-        }
-        lastSurfaceCacheLevel = level;
-        invalidateExteriorFluidChunk(chunkX, chunkZ);
-    }
 
-    public static void invalidateExteriorFluidChunkForTests(final int chunkX, final int chunkZ) {
-        invalidateExteriorFluidChunk(chunkX, chunkZ);
-    }
 
-    public static long getExteriorFluidChunkRevisionForTests(final int chunkX, final int chunkZ) {
-        return currentExteriorFluidChunkRevision(exteriorFluidChunkKeyFromChunk(chunkX, chunkZ));
-    }
 
-    private static void clearExteriorFluidCaches() {
-        EXTERIOR_FLUID_CHUNK_REVISIONS.clear();
-        FLUID_SURFACE_CACHE.clear();
-        FLUID_SURFACE_POINT_CACHE.clear();
-    }
 
+    /**
+     * Emits the overlay quads.
+     *
+     * <p>Must be called from inside the translucent chunk layer, while a shaderpack still has the
+     * terrain-translucent phase active — see {@link #OVERLAY_RENDER_TYPE} for why the call site
+     * decides which program these quads end up being drawn with.</p>
+     */
     public static void render(final double camX, final double camY, final double camZ) {
-        if (!VSGameConfig.CLIENT.getRenderFloodedVoxels()) return;
+        if (!VSGameConfig.CLIENT.getUnderwater().getEnableFluidOverlay()) return;
 
         final Minecraft mc = Minecraft.getInstance();
         final var level = mc.level;
@@ -264,7 +154,7 @@ public final class ShipWaterPocketLiquidOverlay {
         final Camera camera = mc.gameRenderer.getMainCamera();
         if (camera.getFluidInCamera() != FogType.NONE) return;
 
-        ensureExteriorFluidCacheLevel(level);
+        ShipExteriorFluidSampler.ensureExteriorFluidCacheLevel(level);
 
         final Vec3 cameraPos = new Vec3(camX, camY, camZ);
         final Vec3 cameraView = camera.getEntity() != null ? camera.getEntity().getViewVector(1.0F) : new Vec3(0.0, 0.0, 1.0);
@@ -324,22 +214,32 @@ public final class ShipWaterPocketLiquidOverlay {
         bufferSource.endBatch();
     }
 
-    private static RenderType OVERLAY_RENDER_TYPE = new RenderStateShard(null, null, null) {
+    /**
+     * Terrain-translucent state, deliberately, so a shaderpack treats these quads as water.
+     *
+     * <p>Iris picks a program from the render type's shader state and the pipeline's current phase.
+     * Drawing with the terrain translucent shader while the phase is
+     * {@code TERRAIN_TRANSLUCENT} — which is what {@link #render}'s call site guarantees — resolves
+     * to {@code gbuffers_water}, so the pack's own water vertex shader animates and shades this
+     * geometry exactly as it does the world's surface. That also fixes the vertex format:
+     * {@link DefaultVertexFormat#BLOCK} is what terrain programs declare their attributes against,
+     * where the previous entity format would have bound {@code gbuffers_entities} instead.</p>
+     */
+    private static final RenderType OVERLAY_RENDER_TYPE = new RenderStateShard(null, null, null) {
         private static RenderType createOverlayRenderType() {
             return RenderType.create(
                 "valkyrienskies_ship_liquid_overlay",
-                DefaultVertexFormat.NEW_ENTITY,
+                DefaultVertexFormat.BLOCK,
                 VertexFormat.Mode.QUADS,
                 256,
                 false,
                 true,
                 RenderType.CompositeState.builder()
-                    .setShaderState(RenderStateShard.RENDERTYPE_ENTITY_TRANSLUCENT_SHADER)
+                    .setShaderState(RenderStateShard.RENDERTYPE_TRANSLUCENT_SHADER)
                     .setTextureState(new RenderStateShard.TextureStateShard(InventoryMenu.BLOCK_ATLAS, false, false))
                     .setTransparencyState(RenderStateShard.TRANSLUCENT_TRANSPARENCY)
                     .setCullState(RenderStateShard.NO_CULL)
                     .setLightmapState(RenderStateShard.LIGHTMAP)
-                    .setOverlayState(RenderStateShard.OVERLAY)
                     .setLayeringState(RenderStateShard.NO_LAYERING)
                     .setDepthTestState(RenderStateShard.LEQUAL_DEPTH_TEST)
                     .setWriteMaskState(RenderStateShard.COLOR_DEPTH_WRITE)
@@ -632,9 +532,6 @@ public final class ShipWaterPocketLiquidOverlay {
         return net.minecraft.world.level.material.Fluids.WATER;
     }
 
-    static Fluid canonicalSource(final Fluid fluid) {
-        return fluid instanceof net.minecraft.world.level.material.FlowingFluid flowing ? flowing.getSource() : fluid;
-    }
 
     static boolean isOverlaySolidCandidate(final boolean transparentRenderType, final boolean solidRender,
         final boolean propagatesSkylight, final int lightBlock) {
@@ -645,29 +542,6 @@ public final class ShipWaterPocketLiquidOverlay {
         return renderType == RenderType.translucent() || renderType == RenderType.cutout() || renderType == RenderType.cutoutMipped();
     }
 
-    static @Nullable FluidSurfaceSample findExteriorFluidSurface(final net.minecraft.client.multiplayer.ClientLevel level,
-        final BlockPos.MutableBlockPos fluidPos, final BlockPos.MutableBlockPos scanPos,
-        final FluidStateManager.QueryCache fluidQueryCache, final double worldX, final double worldY, final double worldZ) {
-        FluidSurfaceSample best = null;
-        for (final double[] offset : FLUID_SAMPLE_OFFSETS) {
-            final FluidSurfaceSample candidate = findExteriorFluidSurfaceAtSamplePoint(
-                level,
-                fluidPos,
-                scanPos,
-                fluidQueryCache,
-                worldX + offset[0],
-                worldY + offset[1],
-                worldZ + offset[2]
-            );
-            if (candidate == null) {
-                continue;
-            }
-            if (best == null || candidate.surfaceY > best.surfaceY + SURFACE_EPS) {
-                best = candidate;
-            }
-        }
-        return best;
-    }
 
     private static @Nullable OverlayFaceSample resolveOverlayFaceSample(final net.minecraft.client.multiplayer.ClientLevel level,
         final ShipFluidRenderSnapshot snapshot, final BlockPos.MutableBlockPos fluidPos,
@@ -679,7 +553,7 @@ public final class ShipWaterPocketLiquidOverlay {
         final double worldY = m01 * localX + m11 * localY + m21 * localZ + tY;
         final double worldZ = m02 * localX + m12 * localY + m22 * localZ + tZ;
 
-        final FluidSurfaceSample surface = findExteriorFluidSurface(level, fluidPos, scanPos, fluidQueryCache, worldX, worldY, worldZ);
+        final ShipExteriorFluidSampler.FluidSurfaceSample surface = ShipExteriorFluidSampler.findExteriorFluidSurface(level, fluidPos, scanPos, fluidQueryCache, worldX, worldY, worldZ);
         if (surface == null) {
             return null;
         }
@@ -706,232 +580,21 @@ public final class ShipWaterPocketLiquidOverlay {
         );
     }
 
-    private static @Nullable FluidSurfaceSample findExteriorFluidSurfaceAtSamplePoint(
-        final net.minecraft.client.multiplayer.ClientLevel level,
-        final BlockPos.MutableBlockPos fluidPos,
-        final BlockPos.MutableBlockPos scanPos,
-        final FluidStateManager.QueryCache fluidQueryCache,
-        final double worldX,
-        final double worldY,
-        final double worldZ
-    ) {
-        final int blockX = Mth.floor(worldX);
-        final int blockY = Mth.floor(worldY);
-        final int blockZ = Mth.floor(worldZ);
-        final long chunkRevision = currentExteriorFluidChunkRevision(exteriorFluidChunkKey(blockX, blockZ));
-        final long pointKey = BlockPos.asLong(blockX, blockY, blockZ);
-        final ChunkTrackedFluidSurface pointCached = FLUID_SURFACE_POINT_CACHE.get(pointKey);
-        if (pointCached != null) {
-            if (pointCached.chunkRevision == chunkRevision) {
-                return toSample(pointCached.columnHeights, worldX, worldZ);
-            }
-            FLUID_SURFACE_POINT_CACHE.remove(pointKey);
-        }
 
-        fluidPos.set(blockX, blockY, blockZ);
-        FluidState sampleState = getRawExteriorFluidState(level, fluidPos, fluidQueryCache);
-        if (sampleState == null) {
-            fluidPos.move(0, -1, 0);
-            sampleState = getRawExteriorFluidState(level, fluidPos, fluidQueryCache);
-            if (sampleState == null) {
-                fluidPos.move(0, 2, 0);
-                sampleState = getRawExteriorFluidState(level, fluidPos, fluidQueryCache);
-                if (sampleState == null) {
-                    cacheFluidSurfacePoint(pointKey, chunkRevision, COLUMN_HEIGHTS_MISS);
-                    return null;
-                }
-            }
-        }
 
-        final long key = BlockPos.asLong(fluidPos.getX(), fluidPos.getY(), fluidPos.getZ());
-        final ChunkTrackedFluidSurface cached = FLUID_SURFACE_CACHE.get(key);
-        if (cached != null) {
-            if (cached.chunkRevision == chunkRevision) {
-                cacheFluidSurfacePoint(pointKey, chunkRevision, cached.columnHeights);
-                return toSample(cached.columnHeights, worldX, worldZ);
-            }
-            FLUID_SURFACE_CACHE.remove(key);
-        }
 
-        final Fluid canonicalFluid = canonicalSource(sampleState.getType());
-        final int topY = scanRawExteriorFluidColumnTopY(level, fluidPos, canonicalFluid, fluidQueryCache, scanPos);
-        scanPos.set(fluidPos.getX(), topY, fluidPos.getZ());
 
-        final FluidState topFluid = getRawExteriorFluidState(level, scanPos, fluidQueryCache);
-        if (topFluid == null) {
-            cacheFluidSurfacePoint(pointKey, chunkRevision, COLUMN_HEIGHTS_MISS);
-            return null;
-        }
 
-        final ColumnHeights columnHeights = computeVanillaColumnHeights(level, fluidQueryCache, scanPos.immutable(),
-            canonicalFluid, topFluid);
-        if (FLUID_SURFACE_CACHE.size() >= MAX_FLUID_SURFACE_CACHE) {
-            FLUID_SURFACE_CACHE.clear();
-        }
-        FLUID_SURFACE_CACHE.put(key, new ChunkTrackedFluidSurface(chunkRevision, columnHeights));
-        cacheFluidSurfacePoint(pointKey, chunkRevision, columnHeights);
-        return toSample(columnHeights, worldX, worldZ);
-    }
 
-    private static @Nullable FluidSurfaceSample toSample(final ColumnHeights columnHeights, final double worldX, final double worldZ) {
-        if (columnHeights == COLUMN_HEIGHTS_MISS) return null;
-        return new FluidSurfaceSample(columnHeights.fluid, columnHeights.fluidState, columnHeights.pos,
-            columnHeights.surfaceYAt(worldX, worldZ));
-    }
 
-    private static void cacheFluidSurfacePoint(final long pointKey, final long chunkRevision, final ColumnHeights columnHeights) {
-        if (FLUID_SURFACE_POINT_CACHE.size() >= MAX_FLUID_SURFACE_POINT_CACHE) {
-            FLUID_SURFACE_POINT_CACHE.clear();
-        }
-        FLUID_SURFACE_POINT_CACHE.put(pointKey, new ChunkTrackedFluidSurface(chunkRevision, columnHeights));
-    }
 
-    /**
-     * Faithful port of vanilla's {@code LiquidBlockRenderer.calculateAverageHeight} /
-     * {@code addWeightedHeight} / {@code getHeight}: the four corners of a fluid block's top face
-     * are each a weighted average of the block's own height, its two axis-adjacent neighbors, and
-     * (if either axis neighbor has any fluid) the diagonal neighbor — snapping to a flat 1.0 if the
-     * block itself sits under another fluid block of the same type (an interior/non-surface block),
-     * or if either axis neighbor or the diagonal neighbor is itself a full column. This must match
-     * vanilla exactly, not approximate it: any deviation is a real geometric gap between our
-     * occluder/overlay geometry and the actual rendered water mesh, not a precision/bias issue, and
-     * such a gap can go either direction depending on local shoreline/flow topology.
-     */
-    private static ColumnHeights computeVanillaColumnHeights(final net.minecraft.client.multiplayer.ClientLevel level,
-        final FluidStateManager.QueryCache fluidQueryCache, final BlockPos pos, final Fluid canonicalFluid,
-        final FluidState topFluidState) {
-        final float own = vanillaFluidHeight(level, fluidQueryCache, pos, canonicalFluid);
-        if (own >= 1.0f) {
-            return new ColumnHeights(canonicalFluid, topFluidState, pos, 1.0f, 1.0f, 1.0f, 1.0f);
-        }
 
-        final float north = vanillaFluidHeight(level, fluidQueryCache, pos.north(), canonicalFluid);
-        final float south = vanillaFluidHeight(level, fluidQueryCache, pos.south(), canonicalFluid);
-        final float east = vanillaFluidHeight(level, fluidQueryCache, pos.east(), canonicalFluid);
-        final float west = vanillaFluidHeight(level, fluidQueryCache, pos.west(), canonicalFluid);
 
-        final float nw = vanillaCornerHeight(level, fluidQueryCache, canonicalFluid, own, north, west, pos.north().west());
-        final float ne = vanillaCornerHeight(level, fluidQueryCache, canonicalFluid, own, north, east, pos.north().east());
-        final float sw = vanillaCornerHeight(level, fluidQueryCache, canonicalFluid, own, south, west, pos.south().west());
-        final float se = vanillaCornerHeight(level, fluidQueryCache, canonicalFluid, own, south, east, pos.south().east());
-        return new ColumnHeights(canonicalFluid, topFluidState, pos, nw, ne, sw, se);
-    }
 
-    private static float vanillaCornerHeight(final net.minecraft.client.multiplayer.ClientLevel level,
-        final FluidStateManager.QueryCache fluidQueryCache, final Fluid canonicalFluid, final float own,
-        final float axisA, final float axisB, final BlockPos diagonalPos) {
-        if (axisA >= 1.0f || axisB >= 1.0f) {
-            return 1.0f;
-        }
-        final float[] acc = {0.0f, 0.0f};
-        if (axisA > 0.0f || axisB > 0.0f) {
-            final float diag = vanillaFluidHeight(level, fluidQueryCache, diagonalPos, canonicalFluid);
-            if (diag >= 1.0f) {
-                return 1.0f;
-            }
-            addWeightedHeight(acc, diag);
-        }
-        addWeightedHeight(acc, own);
-        addWeightedHeight(acc, axisA);
-        addWeightedHeight(acc, axisB);
-        return acc[1] > 0.0f ? acc[0] / acc[1] : own;
-    }
 
-    private static void addWeightedHeight(final float[] acc, final float h) {
-        if (h >= 0.8f) {
-            acc[0] += h * 10.0f;
-            acc[1] += 10.0f;
-        } else if (h >= 0.0f) {
-            acc[0] += h;
-            acc[1] += 1.0f;
-        }
-        // h < 0.0f means solid (non-fluid, blocking) — contributes nothing, matching vanilla.
-    }
 
-    /** Vanilla's {@code LiquidBlockRenderer.getHeight}: 1.0 if stacked under the same fluid, this
-     * block's own partial height if it's the same fluid but not stacked, 0.0 if open/non-solid and
-     * not fluid, or -1.0 (excluded from the weighted average) if solid.
-     *
-     * <p>Deliberately does <em>not</em> apply the shipyard exclusion {@link #getRawExteriorFluidState}
-     * uses for picking a legitimate starting sample point: vanilla's real renderer has no concept of
-     * shipyards at all, it just reads whatever block is actually there. A neighbor of the sampled
-     * water column falling inside the ship's reserved shipyard bounding region is common right at
-     * the shoreline where the ship meets the water — exactly where this cap connects — and excluding
-     * it (treating it as absent) would pull the weighted average down away from what vanilla actually
-     * rasterizes, reintroducing the same kind of geometric mismatch this method exists to eliminate.</p>
-     */
-    private static float vanillaFluidHeight(final net.minecraft.client.multiplayer.ClientLevel level,
-        final FluidStateManager.QueryCache fluidQueryCache, final BlockPos pos, final Fluid canonicalFluid) {
-        final BlockState state = FluidStateManager.getBlockState(level, pos, fluidQueryCache);
-        final FluidState fs = state.getFluidState();
-        if (fs.getType().isSame(canonicalFluid)) {
-            final BlockState above = FluidStateManager.getBlockState(level, pos.above(), fluidQueryCache);
-            return above.getFluidState().getType().isSame(canonicalFluid) ? 1.0f : fs.getOwnHeight();
-        }
-        return !state.isSolid() ? 0.0f : -1.0f;
-    }
 
-    static int scanRawExteriorFluidColumnTopY(
-        final net.minecraft.client.multiplayer.ClientLevel level,
-        final BlockPos pos,
-        final Fluid canonicalFluid,
-        final FluidStateManager.QueryCache fluidQueryCache,
-        final BlockPos.MutableBlockPos scanPos
-    ) {
-        scanPos.set(pos);
-        final int maxYExclusive = level.getMaxBuildHeight();
-        while (scanPos.getY() < maxYExclusive) {
-            final FluidState current = getRawExteriorFluidState(level, scanPos, fluidQueryCache);
-            if (current == null || canonicalSource(current.getType()) != canonicalFluid) {
-                break;
-            }
-            scanPos.move(0, 1, 0);
-        }
-        return scanPos.getY() - 1;
-    }
 
-    private static @Nullable FluidState getRawExteriorFluidState(
-        final net.minecraft.client.multiplayer.ClientLevel level,
-        final BlockPos pos,
-        final FluidStateManager.QueryCache fluidQueryCache
-    ) {
-        if (VSGameUtilsKt.isBlockInShipyard(level, pos)) {
-            return null;
-        }
-        final FluidState rawFluid = FluidStateManager.getFluidState(level, pos, fluidQueryCache);
-        return shouldUseExteriorFluidSample(false, rawFluid.isEmpty()) ? rawFluid : null;
-    }
-
-    static void ensureExteriorFluidCacheLevel(final net.minecraft.client.multiplayer.ClientLevel level) {
-        if (lastSurfaceCacheLevel == level) return;
-        lastSurfaceCacheLevel = level;
-        clearExteriorFluidCaches();
-    }
-
-    private static void invalidateExteriorFluidChunk(final int chunkX, final int chunkZ) {
-        final long chunkKey = exteriorFluidChunkKeyFromChunk(chunkX, chunkZ);
-        EXTERIOR_FLUID_CHUNK_REVISIONS.put(chunkKey, currentExteriorFluidChunkRevision(chunkKey) + 1L);
-    }
-
-    private static long exteriorFluidChunkKey(final int blockX, final int blockZ) {
-        return ChunkPos.asLong(SectionPos.blockToSectionCoord(blockX), SectionPos.blockToSectionCoord(blockZ));
-    }
-
-    private static long exteriorFluidChunkKeyFromChunk(final int chunkX, final int chunkZ) {
-        return ChunkPos.asLong(chunkX, chunkZ);
-    }
-
-    private static long currentExteriorFluidChunkRevision(final long chunkKey) {
-        return EXTERIOR_FLUID_CHUNK_REVISIONS.get(chunkKey);
-    }
-
-    static boolean shouldUseExteriorFluidSample(final boolean inShipyard, final boolean emptyFluid) {
-        return !inShipyard && !emptyFluid;
-    }
-
-    static float rawExteriorFluidHeight(final FluidState fluidState) {
-        return fluidState.getOwnHeight();
-    }
 
     private static double worldY(final double m01, final double m11, final double m21, final double tY, final float x,
         final float y, final float z) {
@@ -957,12 +620,12 @@ public final class ShipWaterPocketLiquidOverlay {
         final double wy2 = worldY(m01, m11, m21, tY, x, y1, z1);
         final double wy3 = worldY(m01, m11, m21, tY, x, y0, z1);
 
-        if (Double.isNaN(fluidSurfaceY) || Math.max(Math.max(wy0, wy1), Math.max(wy2, wy3)) <= fluidSurfaceY + SURFACE_EPS) {
+        if (Double.isNaN(fluidSurfaceY) || Math.max(Math.max(wy0, wy1), Math.max(wy2, wy3)) <= fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS) {
             quad(consumer, matrix, sprite, x, y0, z0, x, y0, z1, x, y1, z1, x, y1, z0, u0, v0, u1, v1, normalX, 0.0f, 0.0f,
                 r, g, b, a);
             return 1;
         }
-        if (Math.min(Math.min(wy0, wy1), Math.min(wy2, wy3)) > fluidSurfaceY + SURFACE_EPS) return 0;
+        if (Math.min(Math.min(wy0, wy1), Math.min(wy2, wy3)) > fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS) return 0;
 
         CLIP_X0[0] = x;
         CLIP_Y0[0] = y0;
@@ -1008,12 +671,12 @@ public final class ShipWaterPocketLiquidOverlay {
         final double wy2 = worldY(m01, m11, m21, tY, x1, y, z1);
         final double wy3 = worldY(m01, m11, m21, tY, x0, y, z1);
 
-        if (Double.isNaN(fluidSurfaceY) || Math.max(Math.max(wy0, wy1), Math.max(wy2, wy3)) <= fluidSurfaceY + SURFACE_EPS) {
+        if (Double.isNaN(fluidSurfaceY) || Math.max(Math.max(wy0, wy1), Math.max(wy2, wy3)) <= fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS) {
             quad(consumer, matrix, sprite, x0, y, z0, x1, y, z0, x1, y, z1, x0, y, z1, u0, v0, u1, v1, 0.0f, normalY, 0.0f,
                 r, g, b, a);
             return 1;
         }
-        if (Math.min(Math.min(wy0, wy1), Math.min(wy2, wy3)) > fluidSurfaceY + SURFACE_EPS) return 0;
+        if (Math.min(Math.min(wy0, wy1), Math.min(wy2, wy3)) > fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS) return 0;
 
         CLIP_X0[0] = x0;
         CLIP_Y0[0] = y;
@@ -1059,12 +722,12 @@ public final class ShipWaterPocketLiquidOverlay {
         final double wy2 = worldY(m01, m11, m21, tY, x1, y1, z);
         final double wy3 = worldY(m01, m11, m21, tY, x0, y1, z);
 
-        if (Double.isNaN(fluidSurfaceY) || Math.max(Math.max(wy0, wy1), Math.max(wy2, wy3)) <= fluidSurfaceY + SURFACE_EPS) {
+        if (Double.isNaN(fluidSurfaceY) || Math.max(Math.max(wy0, wy1), Math.max(wy2, wy3)) <= fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS) {
             quad(consumer, matrix, sprite, x0, y0, z, x1, y0, z, x1, y1, z, x0, y1, z, u0, v0, u1, v1, 0.0f, 0.0f, normalZ,
                 r, g, b, a);
             return 1;
         }
-        if (Math.min(Math.min(wy0, wy1), Math.min(wy2, wy3)) > fluidSurfaceY + SURFACE_EPS) return 0;
+        if (Math.min(Math.min(wy0, wy1), Math.min(wy2, wy3)) > fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS) return 0;
 
         CLIP_X0[0] = x0;
         CLIP_Y0[0] = y0;
@@ -1295,7 +958,7 @@ public final class ShipWaterPocketLiquidOverlay {
     }
 
     static boolean touchesFace(final double min, final double max, final boolean towardsPositiveAxis) {
-        return towardsPositiveAxis ? max >= 1.0 - SURFACE_EPS : min <= SURFACE_EPS;
+        return towardsPositiveAxis ? max >= 1.0 - ShipExteriorFluidSampler.SURFACE_EPS : min <= ShipExteriorFluidSampler.SURFACE_EPS;
     }
 
     private static int emitSolidAccessibleFaces(final Matrix4f matrix, final VertexConsumer consumer,
@@ -1846,9 +1509,9 @@ public final class ShipWaterPocketLiquidOverlay {
 
     private static boolean isSolidAt(final List<AABB> boxes, final double x, final double y, final double z) {
         for (final AABB box : boxes) {
-            if (x >= box.minX - SURFACE_EPS && x <= box.maxX + SURFACE_EPS &&
-                y >= box.minY - SURFACE_EPS && y <= box.maxY + SURFACE_EPS &&
-                z >= box.minZ - SURFACE_EPS && z <= box.maxZ + SURFACE_EPS) {
+            if (x >= box.minX - ShipExteriorFluidSampler.SURFACE_EPS && x <= box.maxX + ShipExteriorFluidSampler.SURFACE_EPS &&
+                y >= box.minY - ShipExteriorFluidSampler.SURFACE_EPS && y <= box.maxY + ShipExteriorFluidSampler.SURFACE_EPS &&
+                z >= box.minZ - ShipExteriorFluidSampler.SURFACE_EPS && z <= box.maxZ + ShipExteriorFluidSampler.SURFACE_EPS) {
                 return true;
             }
         }
@@ -1905,7 +1568,7 @@ public final class ShipWaterPocketLiquidOverlay {
         }
 
         final AABB bounds = shape.bounds();
-        if (bounds.getXsize() <= SURFACE_EPS || bounds.getYsize() <= SURFACE_EPS || bounds.getZsize() <= SURFACE_EPS) {
+        if (bounds.getXsize() <= ShipExteriorFluidSampler.SURFACE_EPS || bounds.getYsize() <= ShipExteriorFluidSampler.SURFACE_EPS || bounds.getZsize() <= ShipExteriorFluidSampler.SURFACE_EPS) {
             return null;
         }
         return bounds;
@@ -1978,19 +1641,19 @@ public final class ShipWaterPocketLiquidOverlay {
             // Same "stableSample" reuse emitOverlayFaces does: the cell-center sample is tried
             // first and shared across all faces of this outside cell; only recompute per-face if
             // the shared sample failed (e.g. right at the shipyard boundary).
-            final double stableSurfaceY = sampleOverlayFaceSurfaceY(level, fluidPos, scanPos, fluidQueryCache,
+            final double stableSurfaceY = ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos, fluidQueryCache,
                 lx + 0.5, ly + 0.5, lz + 0.5, m00, m10, m20, m01, m11, m21, m02, m12, m22, tX, tY, tZ);
 
             if (lx > 0) {
                 final int n = outsideIdx - 1;
                 if (isInteriorOpen(open, interior, n)) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + FACE_SAMPLE_OFFSET, ly + 0.5, lz + 0.5, m00, m10, m20, m01, m11, m21, m02,
                         m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) visitor.faceX(lx, ly, ly + 1.0f, lz, lz + 1.0f, +1.0f, sy);
                 } else if (isFullCellOverlaySolid(fullCellOverlaySolids, n)
                     || (!open.get(n) && overlaySolids != null && overlaySolids.get(n))) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + FACE_SAMPLE_OFFSET, ly + 0.5, lz + 0.5, m00, m10, m20, m01, m11, m21, m02,
                         m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) {
@@ -2001,13 +1664,13 @@ public final class ShipWaterPocketLiquidOverlay {
             if (lx + 1 < sizeX) {
                 final int n = outsideIdx + 1;
                 if (isInteriorOpen(open, interior, n)) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 1.0 - FACE_SAMPLE_OFFSET, ly + 0.5, lz + 0.5, m00, m10, m20, m01, m11, m21,
                         m02, m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) visitor.faceX(lx + 1, ly, ly + 1.0f, lz, lz + 1.0f, -1.0f, sy);
                 } else if (isFullCellOverlaySolid(fullCellOverlaySolids, n)
                     || (!open.get(n) && overlaySolids != null && overlaySolids.get(n))) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 1.0 - FACE_SAMPLE_OFFSET, ly + 0.5, lz + 0.5, m00, m10, m20, m01, m11, m21,
                         m02, m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) {
@@ -2018,13 +1681,13 @@ public final class ShipWaterPocketLiquidOverlay {
             if (ly > 0) {
                 final int n = outsideIdx - strideY;
                 if (isInteriorOpen(open, interior, n)) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + FACE_SAMPLE_OFFSET, lz + 0.5, m00, m10, m20, m01, m11, m21, m02,
                         m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) visitor.faceY(ly, lx, lx + 1.0f, lz, lz + 1.0f, +1.0f, sy);
                 } else if (isFullCellOverlaySolid(fullCellOverlaySolids, n)
                     || (!open.get(n) && overlaySolids != null && overlaySolids.get(n))) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + FACE_SAMPLE_OFFSET, lz + 0.5, m00, m10, m20, m01, m11, m21, m02,
                         m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) {
@@ -2035,13 +1698,13 @@ public final class ShipWaterPocketLiquidOverlay {
             if (ly + 1 < sizeY) {
                 final int n = outsideIdx + strideY;
                 if (isInteriorOpen(open, interior, n)) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + 1.0 - FACE_SAMPLE_OFFSET, lz + 0.5, m00, m10, m20, m01, m11, m21,
                         m02, m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) visitor.faceY(ly + 1, lx, lx + 1.0f, lz, lz + 1.0f, -1.0f, sy);
                 } else if (isFullCellOverlaySolid(fullCellOverlaySolids, n)
                     || (!open.get(n) && overlaySolids != null && overlaySolids.get(n))) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + 1.0 - FACE_SAMPLE_OFFSET, lz + 0.5, m00, m10, m20, m01, m11, m21,
                         m02, m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) {
@@ -2052,13 +1715,13 @@ public final class ShipWaterPocketLiquidOverlay {
             if (lz > 0) {
                 final int n = outsideIdx - strideZ;
                 if (isInteriorOpen(open, interior, n)) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + 0.5, lz + FACE_SAMPLE_OFFSET, m00, m10, m20, m01, m11, m21, m02,
                         m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) visitor.faceZ(lz, lx, lx + 1.0f, ly, ly + 1.0f, +1.0f, sy);
                 } else if (isFullCellOverlaySolid(fullCellOverlaySolids, n)
                     || (!open.get(n) && overlaySolids != null && overlaySolids.get(n))) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + 0.5, lz + FACE_SAMPLE_OFFSET, m00, m10, m20, m01, m11, m21, m02,
                         m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) {
@@ -2069,13 +1732,13 @@ public final class ShipWaterPocketLiquidOverlay {
             if (lz + 1 < sizeZ) {
                 final int n = outsideIdx + strideZ;
                 if (isInteriorOpen(open, interior, n)) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + 0.5, lz + 1.0 - FACE_SAMPLE_OFFSET, m00, m10, m20, m01, m11, m21,
                         m02, m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) visitor.faceZ(lz + 1, lx, lx + 1.0f, ly, ly + 1.0f, -1.0f, sy);
                 } else if (isFullCellOverlaySolid(fullCellOverlaySolids, n)
                     || (!open.get(n) && overlaySolids != null && overlaySolids.get(n))) {
-                    final double sy = Double.isNaN(stableSurfaceY) ? sampleOverlayFaceSurfaceY(level, fluidPos, scanPos,
+                    final double sy = Double.isNaN(stableSurfaceY) ? ShipExteriorFluidSampler.sampleSurfaceY(level, fluidPos, scanPos,
                         fluidQueryCache, lx + 0.5, ly + 0.5, lz + 1.0 - FACE_SAMPLE_OFFSET, m00, m10, m20, m01, m11, m21,
                         m02, m12, m22, tX, tY, tZ) : stableSurfaceY;
                     if (!Double.isNaN(sy)) {
@@ -2086,19 +1749,6 @@ public final class ShipWaterPocketLiquidOverlay {
         }
     }
 
-    static double sampleOverlayFaceSurfaceY(final ClientLevel level, final BlockPos.MutableBlockPos fluidPos,
-        final BlockPos.MutableBlockPos scanPos, final FluidStateManager.QueryCache fluidQueryCache,
-        final double localX, final double localY, final double localZ,
-        final double m00, final double m10, final double m20,
-        final double m01, final double m11, final double m21,
-        final double m02, final double m12, final double m22,
-        final double tX, final double tY, final double tZ) {
-        final double worldX = m00 * localX + m10 * localY + m20 * localZ + tX;
-        final double worldY = m01 * localX + m11 * localY + m21 * localZ + tY;
-        final double worldZ = m02 * localX + m12 * localY + m22 * localZ + tZ;
-        final FluidSurfaceSample sample = findExteriorFluidSurface(level, fluidPos, scanPos, fluidQueryCache, worldX, worldY, worldZ);
-        return sample != null ? sample.surfaceY : Double.NaN;
-    }
 
     static int clipToSurfaceY(final float[] inX, final float[] inY, final float[] inZ, final float[] inU, final float[] inV,
         final int inCount, final float[] outX, final float[] outY, final float[] outZ, final float[] outU, final float[] outV,
@@ -2111,7 +1761,7 @@ public final class ShipWaterPocketLiquidOverlay {
         float prevU = inU[inCount - 1];
         float prevV = inV[inCount - 1];
         double prevWorldY = worldY(m01, m11, m21, tY, prevX, prevY, prevZ);
-        boolean prevInside = prevWorldY <= fluidSurfaceY + SURFACE_EPS;
+        boolean prevInside = prevWorldY <= fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS;
 
         for (int i = 0; i < inCount; i++) {
             final float curX = inX[i];
@@ -2120,7 +1770,7 @@ public final class ShipWaterPocketLiquidOverlay {
             final float curU = inU[i];
             final float curV = inV[i];
             final double curWorldY = worldY(m01, m11, m21, tY, curX, curY, curZ);
-            final boolean curInside = curWorldY <= fluidSurfaceY + SURFACE_EPS;
+            final boolean curInside = curWorldY <= fluidSurfaceY + ShipExteriorFluidSampler.SURFACE_EPS;
 
             if (prevInside && curInside) {
                 outX[outCount] = curX;
@@ -2235,10 +1885,10 @@ public final class ShipWaterPocketLiquidOverlay {
         final float g, final float b, final float a) {
         final float scaledU = scaleOverlayUv(u);
         final float scaledV = scaleOverlayUv(v);
+        // DefaultVertexFormat.BLOCK: no overlay coords, unlike the entity format this used to use.
         consumer.vertex(matrix, x, y, z)
             .color(r, g, b, a)
             .uv(Mth.lerp(scaledU, sprite.getU0(), sprite.getU1()), Mth.lerp(scaledV, sprite.getV0(), sprite.getV1()))
-            .overlayCoords(OverlayTexture.NO_OVERLAY)
             .uv2(FULL_BRIGHT)
             .normal(nx, ny, nz)
             .endVertex();
