+++ common/src/main/java/org/valkyrienskies/mod/mixin/feature/fluid_render/MixinLevelRenderer.java
@@ -20,6 +20,7 @@ import net.minecraft.world.phys.Vec3;
 import org.valkyrienskies.mod.common.fluid.client.ShipFluidRenderTypes;
 import org.valkyrienskies.mod.common.fluid.client.ShipInteriorFogRenderer;
 import org.valkyrienskies.mod.common.fluid.client.ShipPocketWorldWaterOccluder;
+import org.valkyrienskies.mod.common.fluid.client.ShipWaterPocketLiquidOverlay;
 
 @Mixin(value = LevelRenderer.class, priority = 900)
 public abstract class MixinLevelRenderer {
@@ -63,6 +64,52 @@ public abstract class MixinLevelRenderer {
         }
 
         renderChunkLayer(ShipFluidRenderTypes.AIR_CULL_RENDER_TYPE, poseStack, camX, camY, camZ, projectionMatrix);
+    }
+
+    /**
+     * Emits the ship fluid overlay while the translucent layer is still set up.
+     *
+     * <p>Placement is the whole point: a shaderpack resolves the terrain translucent shader to
+     * {@code gbuffers_water}, but only when the current rendering phase is not entities or block
+     * entities — those resolve to {@code MOVING_BLOCK} instead. Injecting before the layer tears its
+     * state down keeps us inside the terrain-translucent phase, so the pack's water vertex shader
+     * gets to animate and shade this geometry alongside the world's own surface. Drawing it after
+     * the layer, as a standalone pass, would have picked up the wrong program.</p>
+     */
+    @Inject(
+        method = "renderChunkLayer(Lnet/minecraft/client/renderer/RenderType;Lcom/mojang/blaze3d/vertex/PoseStack;DDDLorg/joml/Matrix4f;)V",
+        at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/client/renderer/RenderType;clearRenderState()V",
+            shift = At.Shift.BEFORE
+        ),
+        // Required: silently missing this point would leave the overlay drawn with the wrong
+        // program, or not drawn at all, with nothing to indicate why.
+        require = 1
+    )
+    private void vs$renderShipFluidOverlay(final RenderType renderType, final PoseStack poseStack,
+        final double camX, final double camY, final double camZ, final Matrix4f projectionMatrix,
+        final CallbackInfo ci) {
+        if (renderType != RenderType.translucent()) return;
+        if (this.level == null) return;
+
+        final Matrix4f oldProjection = new Matrix4f(RenderSystem.getProjectionMatrix());
+        final VertexSorting oldVertexSorting = RenderSystem.getVertexSorting();
+
+        final PoseStack modelViewStack = RenderSystem.getModelViewStack();
+        modelViewStack.pushPose();
+        modelViewStack.setIdentity();
+        modelViewStack.mulPoseMatrix(poseStack.last().pose());
+
+        RenderSystem.setProjectionMatrix(projectionMatrix, VertexSorting.DISTANCE_TO_ORIGIN);
+        RenderSystem.applyModelViewMatrix();
+        try {
+            ShipWaterPocketLiquidOverlay.render(camX, camY, camZ);
+        } finally {
+            modelViewStack.popPose();
+            RenderSystem.applyModelViewMatrix();
+            RenderSystem.setProjectionMatrix(oldProjection, oldVertexSorting);
+        }
     }
 
     @Inject(
